requirements.txt
fastapi==0.102.0
uvicorn[standard]==0.23.2
pillow==11.3.0
stability-sdk==0.8.6
python-dotenv==1.1.1
